#6667 feat(wip): Gallery thumbnail view - abstracted img elements from Gallery and Image as ImageElement to allow for common setting of responsive attributes
